+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/yogavredizon/todolist
+
+go 1.21.1
