+++ main.go
@@ -2,90 +2,18 @@ package main
 
 import (
 	"bufio"
-	"errors"
 	"fmt"
 	"os"
 	"strconv"
-	"time"
+
+	rp "github.com/yogavredizon/todolist/pkg/repository"
 )
-
-type Todo struct {
-	Todo []string
-	Time []time.Time
-}
-
-func getHour(hour string) time.Time {
-
-	// if hour
-
-	getHour, _ := time.Parse(time.Kitchen, hour)
-
-	return getHour
-
-}
-
-func (t *Todo) AddTodo(task, timeTodo string) (Todo, error) {
-	if task == "" {
-		return *t, errors.New("task can't blank")
-	}
-
-	for _, todo := range t.Todo {
-		if todo == task {
-			return *t, errors.New("task already taken")
-		}
-	}
-
-	t.Todo = append(t.Todo, task)
-	t.Time = append(t.Time, getHour(timeTodo))
-
-	return *t, nil
-}
-
-func (t *Todo) ShowTodo() {
-
-	for i, todo := range t.Todo {
-		fmt.Printf("%d. %s : %d:%d \n", i+1, todo, t.Time[i].Hour(), t.Time[i].Minute())
-	}
-}
-
-func (t *Todo) RemoveTodo(id int) (int, error) {
-	id = id - 1
-
-	if id > len(t.Todo)-1 {
-		return -1, errors.New("id not found")
-	}
-
-	t.Todo = append(t.Todo[:id], t.Todo[id+1:]...)
-	t.Time = append(t.Time[:id], t.Time[id+1:]...)
-
-	return id + 1, nil
-}
-
-func (t *Todo) UpdateTime(id int, timeTodo string) (int, error) {
-	id -= 1
-	if id > len(t.Time)-1 {
-		return -1, errors.New("id not found")
-	}
-	t.Time[id] = getHour(timeTodo)
-
-	return id + 1, nil
-}
-func (t *Todo) UpdateTodo(id int, todo string) (int, error) {
-	id -= 1
-	if id > len(t.Todo)-1 {
-		return -1, errors.New("id not found")
-	}
-	t.Todo[id] = todo
-
-	return id + 1, nil
-}
 
 func main() {
 
 	run := 1
-	var todo Todo
+	var todo rp.Todo
 	scanner := bufio.NewScanner(os.Stdin)
-	todo.AddTodo("p", ":00PM")
 
 	for run > 0 {
 		fmt.Print("\nTodolist\n")
@@ -152,7 +80,7 @@ func main() {
 			if err != nil {
 				fmt.Printf("\nError : %v \n", err)
 			} else {
-				fmt.Printf("\n %v :Updated \n", res)
+				fmt.Printf("\n Time in  ID %v updated \n", res)
 			}
 		case 4:
 			scanner.Scan()
@@ -164,7 +92,7 @@ func main() {
 			if err != nil {
 				fmt.Printf("\nError %v : \n", err)
 			} else {
-				fmt.Printf("\n %v :Updated \n", res)
+				fmt.Printf("\n Removed ID %v \n", res)
 			}
 		case 5:
 			run = 0
