+++ pkg/model/todo.go
@@ -1,0 +1,8 @@
+package model
+
+import "time"
+
+type Todo struct {
+	Todo []string
+	Time []time.Time
+}
